+++ packages/editor/src/components/editor-container/editor-container.ts
@@ -33,35 +33,19 @@ export class EditorContainer extends LitElement {
   private _placeholderInput!: HTMLInputElement;
 
   @state()
-  placeholderModel = new BlockSchema.page(this.store, {});
+  placeholderModel!: PageBlockModel;
 
   @state()
   unsubscribe = [] as (() => void)[];
 
-  connectedCallback() {
-    super.connectedCallback();
-    this.init();
-  }
-
-  updated() {
-    this.init();
-  }
-
-  private init() {
-    if (!this.store) {
-      throw new Error("EditorContainer's store is not set");
+  update(changedProperties: Map<string, unknown>) {
+    if (changedProperties.has('store')) {
+      this.placeholderModel = new BlockSchema.page(this.store, {});
     }
-
-    this.unsubscribe.forEach(fn => fn());
-    this.subscribeStore();
-
-    // @ts-ignore
-    window.store = this.store;
-    // @ts-ignore
-    window.editor = this;
+    super.update(changedProperties);
   }
 
-  private subscribeStore() {
+  private _subscribeStore() {
     // if undo to empty page, reset to empty placeholder
     const unsubscribeUpdate = this.store.signals.updated.on(() => {
       this.isEmptyPage = this.store.isEmpty;
@@ -82,11 +66,25 @@ export class EditorContainer extends LitElement {
   }
 
   firstUpdated() {
+    if (!this.store) {
+      throw new Error("EditorContainer's store is not set!");
+    }
+
     window.addEventListener('affine.switch-mode', ({ detail }) => {
       this.mode = detail;
     });
 
+    this._subscribeStore();
+
     this._placeholderInput?.focus();
+    // @ts-ignore
+    window.store = this.store;
+    // @ts-ignore
+    window.editor = this;
+  }
+
+  disconnectedCallback() {
+    this.unsubscribe.forEach(fn => fn());
   }
 
   render() {