+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @blocksuite/blocks
 
+## 0.2.19
+
+### Patch Changes
+
+- d8c1df4: refactor: redesign create editor API
+- Updated dependencies [d8c1df4]
+  - @blocksuite/store@0.2.19
+
 ## 0.2.16
 
 ### Patch Changes